context: delete unused label field
It's not necessary to store the label in the context type, instead use
the corresponding library function to get the current in-memory label
container.

The device label information is stored in-memory by libfdisk library
when creating a context or assigning a device.

Avoids possible incosistencies between libfdisk in-memory label and
python-libfdisk context type label information.